ccimx8x: fix ram_size computing on 4GB non-SPL build
4GB RAM is split between 2 memory banks.

The size of both SDRAM banks was being computed only for builds with
SPL.
Extend that function to compute the size of both banks for builds
without SPL too.

Also, ensure the PHYS_SDRAM_1_SIZE macro can fit a 64-bit value so we
can define by default a 4GB size, which requires 33 bits.

https://onedigi.atlassian.net/browse/DEL-9435

+++ include/configs/ccimx8x_common.h
@@ -61,7 +61,7 @@
 /* SDRAM1 size is defined based on the HWID. Set here the default fallback value */
 #define PHYS_SDRAM_1_SIZE		SZ_512M
 #else
-#define PHYS_SDRAM_1_SIZE		(CONFIG_DDR_MB * (unsigned int)SZ_1M)
+#define PHYS_SDRAM_1_SIZE		(CONFIG_DDR_MB * (unsigned long)SZ_1M)
 #endif
 #define PHYS_SDRAM_2			0x880000000
 #define PHYS_SDRAM_2_SIZE		0
